fix: ensure we can checkout PR branches named identically to the base branch
Similar to the issue reported in laminas/laminas-ci-matrix-action#29, we need to ensure that when we fetch the `$GITHUB_REF`, we fetch it into a branch that is named differently than the base branch to avoid collisions and errors.

This patch accomplishes it by adding a `pull/` prefix to the pull request branch name.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>