feat: add Swoole extension to container
This patch compiles and installs the Swoole extension (at version 4.6.7) for PHP versions 7.3, 7.4, and 8.0.
It is NOT enabled by default; enablement is done selectively, as it is done with sqlsrv (combination of a mods-available INI file, and conditional logic when enabling extensions in the entrypoint).
I have built the container locally, and it builds correctly; more importantly, I then ran it against the mezzio-swoole library to ensure I could run unit tests with it, and tests passed.

At some point, we will likely want to:

- Add a script per extension we want to install manually
- Add a script for each of the various functions/conditionals in the entrypoint to make it more modular

For now, however, this works, and will allow us to move forward with adding GHA to mezzio-swoole.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>