Always attempt to pull a newer base image when building docker images
In a previous commit, we introduced `cache-from` and `cache-to`, which lead
to re-using docker image layers as much as possible. This is fine, but it also
means that, on the long term, we may run into stale base layers, where dependencies
do not get updated when the system shifts dramatically.

To prevent that, we do a full re-build when the base image changes: it's an acceptable
cost, for something that changes less frequently, while we still retain caching advantages
when multiple pushes are performed close to each other.