feature: rework `extensions` handling
This will check if an extension is already enabled by the PHP version compiled within this action.
Afterwards, it will check if it is pre-compiled but not yet enabled for the PHP version.
If both checks fail, an extension is marked to be installed.

For PHP 8.1, we do not yet support dynamic installation of extensions as we do for earlier PHP versions. This is due to the fact that PHP 8.1 is not yet provided by the Sury debian repository.

Signed-off-by: Maximilian Bösing <2189546+boesing@users.noreply.github.com>