feat: use pre-built Swoole/OpenSwoole versions
Instead of building Swoole each time we change the container, we can pre-build it once and pull it in, similar to how we consume the sqlsrv extension.
I've created a project, [laminas-ci-swoole-builder](https://github.com/weierophinney/laminas-ci-swoole-builder) that builds both Swoole and OpenSwoole using containers mimicing the Laminas-CI setup.
The artifacts are then packaged and uploaded to an S3 bucket, allowing us to pull them in here.
This can eventually be managed directly by Laminas, but in the meantime, should speed up our container builds, and give us quicker turnaround on new Swoole and OpenSwoole versions.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>