fix: allow installing extensions on PHP 8.1 normally
When releasing 1.14.0, I missed that the extensions script had a special case for 8.1, and discovered it later (https://github.com/mezzio/mezzio-authentication/runs/4397696325?check_suite_focus=true#step:3:17).

This patch removes that case by commenting it out, but leaves the comment as an example of how to vary behavior based on PHP version.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>