Squash `setup/` steps into one single, simple, efficient `apt install` step
Previously, each `/setup` script was run sequentially, in a separate `RUN` step of the `Dockerfile`.

While that removes some cache slamming problems when developing the image, it is much more efficient
to run the whole `apt install` operation in a single docker layer, therefore removing any potential
layer shifting that leads to many more layers, and much larger final docker image.

In practice, we only expect to change this section rarely, and relying on the built-in `apt` sources
is sufficient, for now.

For comparison, `ghcr.io/laminas/laminas-continuous-integration:1.15.0` is `1.4Gb` in size,
while the new image is `1.1Gb`, shaving off 300Mb of unused dependencies.