feat: lazy-install swoole, openswoole extensions
Instead of shipping them in the container, this patch modifies the workflow to instead detect when swoole or openswoole have been requested within the entrypoint, and then downloads a package for the given extension specific to the current PHP version.
The package is extracted, and then `phpenmod` used to enable it.

The goal here is to (a) reduce overall container size, (b) reduce build times for this container, and (c) result in faster container startup times.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>