feature: handle PDO driver extensions
Projects `composer.json` can require specific PDO drivers such as `ext-pdo_mysql`. These are passed to the CI container as `pdo_mysql` and are then passed as `php<version>-pdo_mysql` package to `apt install`. These packages do not exist and thus there will be an error if the appropriate extension(s) was/were not already installed/enabled.
This patch handles the most common drivers `mysql`, `pgsql` and `sqlite` and ensure that these are installed/enabled along with the `pdo` extension.

Signed-off-by: Maximilian Bösing <2189546+boesing@users.noreply.github.com>